Add comments for readability of code
